+++ src/main/java/com/qsp/movie_reservation_system/service/FoodService.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Service;
 
 import com.qsp.movie_reservation_system.dao.FoodDao;
 import com.qsp.movie_reservation_system.dto.Food;
+import com.qsp.movie_reservation_system.exceptions.FoodIdNotFound;
 import com.qsp.movie_reservation_system.util.ResponseStructure;
 import com.qsp.movie_reservation_system.util.ResponseStructure1;
 
@@ -30,19 +31,24 @@ public class FoodService {
 	}
 
 	public ResponseStructure<Food> fetchFoodById(int foodId) {
-		
+		Food food=foodDao.fetchFoodById(foodId);
+		if(food != null) {
 		responseStructure.setStatusCode(HttpStatus.FOUND.value());
 		responseStructure.setMessage("Successfully Food Fetched from Database");
         responseStructure.setData(foodDao.fetchFoodById(foodId));
 		return responseStructure;
-
+		}
+		throw new FoodIdNotFound(); 
 	}
 	public ResponseStructure<Food> updateFoodById(int oldFoodId,Food newFood) {
-		
+		Food food=foodDao.fetchFoodById(oldFoodId);
+		if(food != null) {
 		responseStructure.setStatusCode(HttpStatus.OK.value());
 		responseStructure.setMessage("Successfully Food Updated from Database");
 		responseStructure.setData(foodDao.updateFoodById(oldFoodId, newFood));
 		return responseStructure ;
+	}
+	throw new FoodIdNotFound(); 
 	}
 	
 	public ResponseStructure1<Food> fetchAllFood(){
@@ -56,12 +62,15 @@ public class FoodService {
 	}
 
 	public ResponseStructure<Food> deleteFoodById(int foodId){
-		
+		Food food=foodDao.fetchFoodById(foodId);
+		if(food != null) {
 		responseStructure.setStatusCode(HttpStatus.OK.value());
 		responseStructure.setMessage("Successfully Food delete from Database");
 		responseStructure.setData(foodDao.deleteFoodById(foodId));
 		return responseStructure ;
-		
+		}
+		throw new FoodIdNotFound(); 
 	}
+		
 	
 }
