+++ src/main/java/com/qsp/movie_reservation_system/service/SeatService.java
@@ -5,6 +5,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import com.qsp.movie_reservation_system.dao.SeatDao;
 import com.qsp.movie_reservation_system.dto.Seat;
+import com.qsp.movie_reservation_system.exceptions.SeatIdNotFound;
 import com.qsp.movie_reservation_system.util.ResponseStructure;
 import com.qsp.movie_reservation_system.util.ResponseStructure1;
 
@@ -28,19 +29,26 @@ public class SeatService {
 	}
 
 	public ResponseStructure<Seat> fetchSeatById(int seatId) {
-		
+		Seat dbSeat=seatDao.fetchSeatById(seatId);
+		if(dbSeat != null) {
 		responseStructure.setStatusCode(HttpStatus.FOUND.value());
 		responseStructure.setMessage("Successfully Seat Fetched from Database");
         responseStructure.setData(seatDao.fetchSeatById(seatId));
 		return responseStructure;
+		
+		}
+		throw new SeatIdNotFound();
 
 	}
 	public ResponseStructure<Seat> updateSeatById(int oldSeatId,Seat newSeat) {
-		
+		Seat dbSeat=seatDao.fetchSeatById(oldSeatId);
+		if(dbSeat != null) {
 		responseStructure.setStatusCode(HttpStatus.OK.value());
 		responseStructure.setMessage("Successfully Seat Updated from Database");
 		responseStructure.setData(seatDao.updateSeatById(oldSeatId, newSeat));
 		return responseStructure ;
+		}
+		throw new SeatIdNotFound();
 	}
 	
 	public ResponseStructure1<Seat> fetchAllSeat(){
@@ -52,12 +60,14 @@ public class SeatService {
 	}
 
 	public ResponseStructure<Seat> deleteSeatById(int seatId){
-		
+		Seat dbSeat=seatDao.fetchSeatById(seatId);
+		if(dbSeat != null) {
 		responseStructure.setStatusCode(HttpStatus.OK.value());
 		responseStructure.setMessage("Successfully Seat delete from Database");
 		responseStructure.setData(seatDao.deleteSeatById(seatId));
 		return responseStructure ;
-		
+		}
+		throw new SeatIdNotFound();
 	}
